extend to accept Vendor/Product and Name
This extends usbreset to not only accept to specify the USB device with
bus and device number, but also via the vendor and product id or product
name. This is accomplished by searching sysfs to find the device based on
the given specifier.
In contrast to the previous version, this also has the effect that bus
and device numbers are now properly checked via sysfs.

Signed-off-by: Jonas Jelonek <jonas.jelonek@protonmail.ch>